+++ advanced-responsive-video-embedder/php/Video.php
@@ -134,6 +134,7 @@ class Video {
 
 			check_product_keys();
 			$this->process_shortcode_atts();
+			$this->oembed_data_errors();
 
 			$html .= get_error_html();
 			$html .= $this->build_html();
@@ -164,6 +165,24 @@ class Video {
 		return apply_filters( 'nextgenthemes/arve/html', $html, get_object_vars( $this ) );
 	}
 
+	private function oembed_data_errors(): void {
+
+		if ( isset( $this->oembed_data->youtube_api_error )
+			&& str_contains( $this->oembed_data->youtube_api_error, '403' )
+		) {
+			update_option( 'arve_youtube_api_error', $this->oembed_data->youtube_api_error );
+			unset( $this->oembed_data->youtube_api_error );
+		}
+
+		unset( $this->oembed_data->arve_error ); // ignore old errors.
+
+		foreach ( (array) $this->oembed_data as $key => $value ) {
+			if ( str_contains( $key, 'error' ) ) {
+				arve_errors()->add( $key, $value );
+			}
+		}
+	}
+
 	private function process_shortcode_atts(): void {
 
 		$this->missing_attribute_check();
@@ -171,10 +190,6 @@ class Video {
 
 		foreach ( $this->shortcode_atts as $arg_name => $value ) {
 			$this->set_prop( $arg_name, $value );
-		}
-
-		if ( ! empty( $this->oembed_data->arve_error ) ) {
-			arve_errors()->add( 'oembed-data-error', $this->oembed_data->arve_error );
 		}
 
 		if ( ! empty( $this->oembed_data->provider ) &&