+++ advanced-responsive-video-embedder/changelog.md
@@ -3,6 +3,17 @@
 * [ARVE Pro changelog](https://nextgenthemes.com/plugins/arve-pro/#changelog)
 * [ARVE Random Videos changelog](https://nextgenthemes.com/plugins/arve-random-video/#changelog)
 
+### 2025-03-09 10.6.4 ###
+
+* Fix: Link to settings page in API key notice.
+* Improved: Link to tutorial video showing you how to get an YouTube Data API key.
+
+### 2025-03-09 10.6.1 ###
+
+* Improved: Error handling. Show YouTube API errors only on the Admin screens.
+* Improved: Cache deletion.
+* New: Option to add your own YouTube Data API key to prevent limits of the one included in Pro.
+
 ### 2025-03-08 10.6.0 ###
 
 * New: Lazyload Style option (Pro Feature).